requirements.txt
python-dotenv
python_gitlab
retry
typer
requests
